moving LogRetroCommand into its own file
