+++ src/commands/log/LogRetroCommand.js
@@ -1,12 +1,8 @@
-import errorReporter from '../util/errorReporter'
-import graphQLFetcher from '../util/graphQLFetcher'
-import getServiceBaseURL, {GAME} from '../util/getServiceBaseURL'
-import loadCommand from '../util/loadCommand'
-import composeInvoke from '../util/composeInvoke'
+import errorReporter from '../../util/errorReporter'
+import graphQLFetcher from '../../util/graphQLFetcher'
+import getServiceBaseURL, {GAME} from '../../util/getServiceBaseURL'
 
-export const {parse, usage, commandDescriptor} = loadCommand('log')
-
-class LogRetroCommand {
+export default class LogRetroCommand {
   constructor(lgJWT, notify, formatMessage, formatError) {
     this.runGraphQLQuery = graphQLFetcher(lgJWT, getServiceBaseURL(GAME))
     this.notifyMsg = msg => notify(formatMessage(msg))
@@ -114,29 +110,3 @@ class LogRetroCommand {
       })
   }
 }
-
-export const invoke = composeInvoke(parse, usage, (args, notify, options) => {
-  const {
-    lgJWT,
-    lgPlayer,
-    formatError,
-    formatMessage,
-  } = options
-  if (!lgJWT || !lgPlayer || !lgPlayer.id) {
-    return Promise.reject('You are not a player in the game.')
-  }
-  if (args.retro) {
-    const retro = new LogRetroCommand(lgJWT, notify, formatMessage, formatError)
-
-    if (typeof args.question === 'string' && args.question.match(/^\d+$/)) {
-      const questionNumber = parseInt(args.question, 10)
-      const responseParams = args._
-      if (responseParams.length === 0) {
-        return retro.printSurveyQuestion(questionNumber)
-      }
-      return retro.logReflection(questionNumber, responseParams)
-    }
-    return retro.printSurvey()
-  }
-  return Promise.reject('Invalid arguments. Try --help for usage.')
-})
